requirements.txt
discord.py
aiohttp
requests
websockets
chardet
